+++ src/main/java/com/example/springbootapi/dto/PaymentsDTO.java
@@ -2,6 +2,7 @@ package com.example.springbootapi.dto;
 
 import lombok.Data;
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
 @Data
@@ -9,12 +10,11 @@ public class PaymentsDTO {
     private Integer id;
     private Integer orderId;
     private String paymentMethod;
-    private Double amount;
+    private BigDecimal amount;
     private String paymentStatus;
     private LocalDateTime paymentDate;
 
-    // Constructor thủ công
-    public PaymentsDTO(Integer id, Integer orderId, String paymentMethod, Double amount, String paymentStatus, LocalDateTime paymentDate) {
+    public PaymentsDTO(Integer id, Integer orderId, String paymentMethod, BigDecimal amount, String paymentStatus, LocalDateTime paymentDate) {
         this.id = id;
         this.orderId = orderId;
         this.paymentMethod = paymentMethod;
